+++ src/react/src/reducers/index.js
@@ -38,6 +38,7 @@ import ContributeProductionLocationReducer from './ContributeProductionLocationR
 import DashboardModerationQueueReducer from './DashboardModerationQueueReducer';
 import DashboardContributionRecordReducer from './DashboardContributionRecordReducer';
 import ClaimFormReducer from './ClaimFormReducer';
+import ClaimCampaignsReducer from './ClaimCampaignsReducer';
 import PartnerFieldGroupsReducer from './PartnerFieldGroupsReducer';
 import PartnerGroupContributorsReducer from './PartnerGroupContributorsReducer';
 import SectionNavigationReducer from './SectionNavigationReducer';
@@ -75,6 +76,7 @@ export default combineReducers({
     dashboardModerationQueue: DashboardModerationQueueReducer,
     dashboardContributionRecord: DashboardContributionRecordReducer,
     claimForm: ClaimFormReducer,
+    claimCampaigns: ClaimCampaignsReducer,
     partnerFieldGroups: PartnerFieldGroupsReducer,
     partnerGroupContributors: PartnerGroupContributorsReducer,
     sectionNavigation: SectionNavigationReducer,